--- a/buildpatches/com_github_valyala_gozstd.patch
+++ b/buildpatches/com_github_valyala_gozstd.patch
@@ -1,0 +1,127 @@
+diff --git a/BUILD.bazel b/BUILD.bazel
+index 2f968a0..e90eb5d 100644
+--- a/BUILD.bazel
++++ b/BUILD.bazel
+@@ -1,69 +1,79 @@
+ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+ 
+-go_library(
+-    name = "gozstd",
+-    srcs = [
+-        "dict.go",
+-        "doc.go",
+-        "gozstd.go",
+-        "libzstd_darwin_amd64.go",
+-        "libzstd_darwin_arm64.go",
+-        "libzstd_freebsd_amd64.go",
+-        "libzstd_illumos_amd64.go",
+-        "libzstd_linux_amd64.go",
+-        "libzstd_linux_arm.go",
+-        "libzstd_linux_arm64.go",
+-        "libzstd_windows_amd64.go",
+-        "reader.go",
+-        "stream.go",
+-        "writer.go",
+-        "zdict.h",
+-        "zstd.h",
+-        "zstd_errors.h",
+-    ],
+-    cgo = True,
+-    clinkopts = select({
++cc_library(
++    name = "libzstd",
++    srcs = select({
+         "@io_bazel_rules_go//go/platform:android_amd64": [
+-            "./libzstd_linux_amd64.a",
++            "libzstd_linux_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:android_arm": [
+-            "./libzstd_linux_arm.a",
++            "libzstd_linux_arm.a",
+         ],
+         "@io_bazel_rules_go//go/platform:android_arm64": [
+-            "./libzstd_linux_arm64.a",
++            "libzstd_linux_arm64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:darwin_amd64": [
+-            "./libzstd_darwin_amd64.a",
++            "libzstd_darwin_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:darwin_arm64": [
+-            "./libzstd_darwin_arm64.a",
++            "libzstd_darwin_arm64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:freebsd_amd64": [
+-            "./libzstd_freebsd_amd64.a",
++            "libzstd_freebsd_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:illumos_amd64": [
+-            "./libzstd_illumos_amd64.a",
++            "libzstd_illumos_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:ios_amd64": [
+-            "./libzstd_darwin_amd64.a",
++            "libzstd_darwin_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:ios_arm64": [
+-            "./libzstd_darwin_arm64.a",
++            "libzstd_darwin_arm64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:linux_amd64": [
+-            "./libzstd_linux_amd64.a",
++            "libzstd_linux_amd64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:linux_arm": [
+-            "./libzstd_linux_arm.a",
++            "libzstd_linux_arm.a",
+         ],
+         "@io_bazel_rules_go//go/platform:linux_arm64": [
+-            "./libzstd_linux_arm64.a",
++            "libzstd_linux_arm64.a",
+         ],
+         "@io_bazel_rules_go//go/platform:windows_amd64": [
+-            "./libzstd_windows_amd64.a",
++            "libzstd_windows_amd64.a",
+         ],
+-        "//conditions:default": [],
++        "//conditions:default": ["UNSUPPORTED_PLATFORM"],
+     }),
++    hdrs = [
++        "zdict.h",
++        "zstd.h",
++        "zstd_errors.h",
++    ],
++)
++
++go_library(
++    name = "gozstd",
++    srcs = [
++        "dict.go",
++        "doc.go",
++        "gozstd.go",
++        "libzstd_darwin_amd64.go",
++        "libzstd_darwin_arm64.go",
++        "libzstd_freebsd_amd64.go",
++        "libzstd_illumos_amd64.go",
++        "libzstd_linux_amd64.go",
++        "libzstd_linux_arm.go",
++        "libzstd_linux_arm64.go",
++        "libzstd_windows_amd64.go",
++        "reader.go",
++        "stream.go",
++        "writer.go",
++        "zdict.h",
++        "zstd.h",
++        "zstd_errors.h",
++    ],
++    cdeps = [":libzstd"],
++    cgo = True,
+     copts = ["-O3"],
+     importpath = "github.com/valyala/gozstd",
+     visibility = ["//visibility:public"],
+@@ -88,3 +98,5 @@ go_test(
+     ],
+     embed = [":gozstd"],
+ )
++
++# gazelle:prefix github.com/valyala/gozstd