--- a/deps.bzl
+++ b/deps.bzl
@@ -5146,8 +5146,9 @@ def install_go_mod_dependencies(workspace_name = "buildbuddy"):
     )
     go_repository(
         name = "com_github_valyala_gozstd",
-		build_file_generation = "on",
         importpath = "github.com/valyala/gozstd",
+        patch_args = ["-p1"],
+        patches = ["//buildpatches:com_github_valyala_gozstd.patch"],
         sum = "h1:xPnnnvjmaDDitMFfDxmQ4vpx0+3CdTg2o3lALvXTU/g=",
         version = "v1.20.1",
     )