--- a/server/util/compression/BUILD
+++ b/server/util/compression/BUILD
@@ -8,8 +8,8 @@ go_library(
     deps = [
         "//server/metrics",
         "//server/util/log",
-        "@com_github_klauspost_compress//zstd",
         "@com_github_prometheus_client_golang//prometheus",
+        "@com_github_valyala_gozstd//:go_default_library",
     ],
 )