@luluz66 2024-01-10
The following result is running pebble benchmark with background go routines disabled. 
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                          │ /home/lulu/benchmark-zstd-before.log │ /home/lulu/benchmark-zstd-after.log │
                          │                sec/op                │    sec/op     vs base               │
GetMulti/size=1KiB-24                               3.143m ± 10%    2.797m ± 8%  -10.99% (p=0.014 n=9)
GetMulti/size=1MiB-24                               77.15m ±  3%    97.07m ± 2%  +25.82% (p=0.000 n=9)
GetMulti/size=10MiB-24                              832.9m ±  1%    799.6m ± 2%   -4.00% (p=0.000 n=9)
FindMissing/size=1KiB-24                            723.4µ ± 22%    802.8µ ± 3%  +10.98% (p=0.000 n=9)
FindMissing/size=1MiB-24                            6.914m ±  2%    9.336m ± 2%  +35.03% (p=0.000 n=9)
FindMissing/size=10MiB-24                           47.55m ±  1%    69.94m ± 2%  +47.10% (p=0.000 n=9)
Contains1/size=1KiB-24                              186.8µ ± 39%    160.1µ ± 3%        ~ (p=0.050 n=9)
Contains1/size=1MiB-24                              6.963m ±  6%    8.664m ± 2%  +24.42% (p=0.000 n=9)
Contains1/size=10MiB-24                             47.13m ±  1%    69.82m ± 5%  +48.13% (p=0.000 n=9)
Set/size=1KiB-24                                    169.4µ ±  4%    146.0µ ± 1%  -13.81% (p=0.004 n=9)
Set/size=1MiB-24                                    4.256m ±  4%    4.296m ± 1%        ~ (p=0.666 n=9)
Set/size=10MiB-24                                   27.95m ±  2%    31.34m ± 1%  +12.11% (p=0.000 n=9)
geomean                                             7.903m          8.804m       +11.40%

                          │ /home/lulu/benchmark-zstd-before.log │   /home/lulu/benchmark-zstd-after.log   │
                          │                 B/op                 │      B/op       vs base                 │
GetMulti/size=1KiB-24                              821.7Ki ±  0%    656.3Ki ±  0%    -20.13% (p=0.000 n=9)
GetMulti/size=1MiB-24                              115.1Mi ±  0%    114.4Mi ±  0%     -0.64% (p=0.000 n=9)
GetMulti/size=10MiB-24                             1.102Gi ±  0%    1.093Gi ±  0%     -0.74% (p=0.000 n=9)
FindMissing/size=1KiB-24                           294.5Ki ±  0%    298.1Ki ±  0%     +1.23% (p=0.000 n=9)
FindMissing/size=1MiB-24                           9.300Mi ±  0%    9.878Mi ±  1%     +6.22% (p=0.000 n=9)
FindMissing/size=10MiB-24                          82.33Mi ±  0%    86.20Mi ±  0%     +4.70% (p=0.000 n=9)
Contains1/size=1KiB-24                             29.94Ki ±  2%    29.98Ki ±  0%          ~ (p=0.422 n=9)
Contains1/size=1MiB-24                             9.031Mi ±  0%    9.636Mi ±  0%     +6.69% (p=0.000 n=9)
Contains1/size=10MiB-24                            82.11Mi ±  0%    85.94Mi ±  0%     +4.66% (p=0.000 n=9)
Set/size=1KiB-24                                   11.82Ki ±  4%    11.85Ki ±  4%          ~ (p=0.473 n=9)
Set/size=1MiB-24                                   33.77Ki ± 32%   779.17Ki ± 21%  +2207.04% (p=0.000 n=9)
Set/size=10MiB-24                                  471.9Ki ± 26%   4182.4Ki ±  0%   +786.21% (p=0.000 n=9)
geomean                                            2.581Mi          4.019Mi          +55.71%

                          │ /home/lulu/benchmark-zstd-before.log │ /home/lulu/benchmark-zstd-after.log  │
                          │              allocs/op               │  allocs/op   vs base                 │
GetMulti/size=1KiB-24                               12.098k ± 0%   9.301k ± 0%  -23.12% (p=0.000 n=9)
GetMulti/size=1MiB-24                                17.43k ± 0%   10.30k ± 0%  -40.89% (p=0.000 n=9)
GetMulti/size=10MiB-24                               75.70k ± 0%   19.43k ± 0%  -74.33% (p=0.000 n=9)
FindMissing/size=1KiB-24                             6.190k ± 0%   6.191k ± 0%   +0.02% (p=0.000 n=9)
FindMissing/size=1MiB-24                             6.226k ± 0%   6.250k ± 0%   +0.39% (p=0.000 n=9)
FindMissing/size=10MiB-24                            6.245k ± 0%   6.285k ± 0%   +0.64% (p=0.000 n=9)
Contains1/size=1KiB-24                                249.0 ± 0%    249.0 ± 0%        ~ (p=1.000 n=9) ¹
Contains1/size=1MiB-24                                285.0 ± 0%    309.0 ± 0%   +8.42% (p=0.000 n=9)
Contains1/size=10MiB-24                               304.0 ± 0%    344.0 ± 0%  +13.16% (p=0.000 n=9)
Set/size=1KiB-24                                      164.0 ± 0%    164.0 ± 0%        ~ (p=1.000 n=9) ¹
Set/size=1MiB-24                                      165.0 ± 1%    191.0 ± 2%  +15.76% (p=0.000 n=9)
Set/size=10MiB-24                                     171.0 ± 1%    200.0 ± 1%  +16.96% (p=0.000 n=9)
geomean                                              1.642k        1.433k       -12.71%
¹ all samples are equal
```