@luluz66 2024-01-11
It's a bit wierd that FindMissing and Contains are slower when it don't involve compression/decompression. I checked the cpu profiles during the benchmark and it shows the garbage collector is more invovled with valyala library. In the benchmark test, we do call pebble_cache.Set to set up the benchmark. Even though we reset the timer after the setup, this step might allocate different amount of space causing different behavior during the actual test.