@bduffany 2024-01-11
> Even though we reset the timer after the setup, this step might allocate different amount of space causing different behavior during the actual test.

just a thought - what if we call `runtime.GC()` just before resetting the timer?